add some test code

+++ oauth2-authorization-server/src/main/java/com/example/oauth2authorizationserver/entity/TestEnum2.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
 public enum  TestEnum2 {
-    TT("desc1",(short)3), TT2("desc2",(short) 4);
+    TT("desc1",3), TT2("desc2",4);
 
     private String desc;
-    private short index;
+    private int index;
 
 
-    TestEnum2(String desc2, short i) {
+    TestEnum2(String desc2, int i) {
         this.desc = desc2;
         this.index = i;
     }
@@ -23,15 +23,15 @@ public enum  TestEnum2 {
         this.desc = desc;
     }
 
-    public short getIndex() {
+    public int getIndex() {
         return index;
     }
 
-    public void setIndex(short index) {
+    public void setIndex(int index) {
         this.index = index;
     }
 
-    public static TestEnum2 valueOf(Short dbData) {
+    public static TestEnum2 valueOf(Integer dbData) {
         switch (dbData){
             case 3:
                 return TT;
