+++ oauth2-authorization-server/src/main/java/com/example/oauth2authorizationserver/entity/TestEnum2Converter.java
@@ -2,14 +2,14 @@ package com.example.oauth2authorizationserver.entity;
 
 import javax.persistence.AttributeConverter;
 
-public class TestEnum2Converter implements AttributeConverter<TestEnum2, Short> {
+public class TestEnum2Converter implements AttributeConverter<TestEnum2, Integer> {
     @Override
-    public Short convertToDatabaseColumn(TestEnum2 attribute) {
+    public Integer convertToDatabaseColumn(TestEnum2 attribute) {
         return attribute.getIndex();
     }
 
     @Override
-    public TestEnum2 convertToEntityAttribute(Short dbData) {
+    public TestEnum2 convertToEntityAttribute(Integer dbData) {
         return TestEnum2.valueOf(dbData);
     }
 }
